requirements.txt
python-dotenv==1.0.1
praw==7.7.1
pydantic==2.8.2
PyYAML==6.0.2
tqdm==4.66.4
python-slugify==8.0.4
requests==2.32.3
